Extract shared graphlink_webengine.py hardening module
Both the HTML Renderer node's preview and the composer's React island render
content in a QWebEngineView and need the same guarantee: nothing they show may
reach the network or local disk. That guarantee - an off-the-record profile, a
local-only request interceptor, and defense-in-depth settings - lived entirely
inside graphlink_html_view.py, with the composer reaching into that file to
reuse it. It now lives in its own module, graphlink_webengine.py, and both
consumers import from it directly.

An independent adversarial review of the diff (fresh context, asked to disprove
"zero behavior change") found a real gap: splitting graphlink_html_view.py's
WebEngine import block let WEBENGINE_AVAILABLE and QWebEngineView's actual
bindability disagree under a partial PySide6 install (QtWebEngineCore
importable, QtWebEngineWidgets not) - WEBENGINE_AVAILABLE could read True while
QWebEngineView stayed unbound, crashing HtmlViewNode with NameError instead of
falling back to its placeholder UI. Fixed by ANDing the shared module's
availability with the node's own local import success, and added a regression
test - without reloading graphlink_html_view itself, since 11 other modules
cache HtmlViewNode at their own import time and reloading would fork class
identity across the shared-process test session.

Verified: python -m compileall passes, full suite 487 passed (486 existing +
1 new), singleton profile/interceptor sharing across both consumers confirmed
by identity check.
